# docs/swagger.yaml
basePath: /v1
definitions:
...
  routes.MoneyAmount:
    properties:
      amount:
        description: always a string — never float for money
        type: string
      currency:
        description: '"NGN"'
        type: string
    type: object
...
  routes.TransferResponse:
    properties:
      amount:
        $ref: '#/definitions/routes.MoneyAmount'
      client_reference:
        type: string
      description:
        type: string
      destination_account_id:
        type: string
      external_reference:
        type: string
      failure_code:
        type: string
      failure_reason:
        type: string
      fee:
        $ref: '#/definitions/routes.MoneyAmount'
      id:
        type: string
      journal_transaction_id:
        type: string
      posted_at:
        type: string
      requested_at:
        type: string
      source_account_id:
        type: string
      status:
        type: string
    type: object
...
  routes.WithddrawParams:
    properties:
      amount:
        type: string
      client_reference:
        type: string
      currency_code:
        type: string
      customer_id:
        type: string
      description:
        type: string
      destination_account_id:
        type: string
      external_reference:
        type: string
      fee_amount:
        type: string
      idempotency_key_id:
        type: string
      memo:
        type: string
      metadata:
        additionalProperties: true
        type: object
      source_account_id:
        type: string
      source_system:
        type: string
    type: object
  routes.errResponse:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: A comprehensive distributed payment processing system with support
    for transfers, deposits, withdrawals, and holds.
  title: Distributed Payments Engine API
  version: 1.0.0
paths:
...
  /account/withdraw:
    post:
      consumes:
      - application/json
      description: Debits funds from an account to the system settlement account.
        Supports idempotent withdrawals.
      parameters:
      - description: Withdrawal Details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/routes.WithddrawParams'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/routes.TransferResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.errResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/routes.errResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.errResponse'
      summary: Withdraw funds from account
      tags:
      - Withdrawals
...
swagger: "2.0"